Remove previous project dates field definitions
